+++ boluo-core/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - 将 `around_with_state_fn` 和 `around_fn` 分别重命名为 `filter_fn_with_state` 和 `filter_fn`，并重构相关代码。
 - 避免多个 `OnUpgrade` 重复调用导致恐慌，并重构相关代码。
+- 使用 AsyncFn* 重构异步闭包约束，使闭包可以返回借用其捕获值的 Future。
 
 ## 新增